Revert "Adding suggested friends to the server ..."
