+++ Client/src/Components/Home/Widgets/Widgets.js
@@ -1,40 +1,101 @@
-import React, { useState, useEffect } from "react";
-import { Link } from "react-router-dom";
+import React from "react";
 import "./Widgets.css";
-import PeopleBox from '../../PeopleBox/PeopleBox';
-import getTokenConfig from '../../../Utils/TokenUtils';
-import axios from "axios";
-
 
 const Widgets = () => {
-    const [users, setUsers] = useState([]);
-    
-    useEffect(() => {
-        const fetchUsers = async () => {
-            try {
-                const config = getTokenConfig();
-                if (!config) return;
-                const response = await axios.get(`http://localhost:3003/api/users`, config);
-                setUsers(response.data.data);
-            } catch (error) {
-                console.error("Error fetching posts:", error);
-            }
-        };
-
-        fetchUsers(); 
-    }, []);
-    
     return (
         <div className="widgets">
             <div className="title">You might like</div>
+            <div className="suggestFriend">
+                <div className="suggestImgContainer">
+                    <img className="suggestImg" src="/images" alt="" />
+                </div>
+                <div className="name">
+                    <span className="suggestName">Friend Name  <br /> <span className="handel">@Handel</span> </span>
+                </div>
+                <button> follow</button>
+            </div>
+            <div className="suggestFriend">
+                <div className="suggestImgContainer">
+                    <img className="suggestImg" src="/images" alt="" />
+                </div>
+                <div className="name">
+                    <span className="suggestName">Friend Name  <br /> <span className="handel">@Handel</span> </span>
+                </div>
+                <button> follow</button>
+            </div>
+            <div className="suggestFriend">
+                <div className="suggestImgContainer">
+                    <img className="suggestImg" src="/images" alt="" />
+                </div>
+                <div className="name">
+                    <span className="suggestName">Friend Name  <br /> <span className="handel">@Handel</span> </span>
+                </div>
+                <button> follow</button>
+            </div>
+            <div className="suggestFriend">
+                <div className="suggestImgContainer">
+                    <img className="suggestImg" src="/images" alt="" />
+                </div>
+                <div className="name">
+                    <span className="suggestName">Friend Name  <br /> <span className="handel">@Handel</span> </span>
+                </div>
+                <button> follow</button>
+            </div>
+            <div className="suggestFriend">
+                <div className="suggestImgContainer">
+                    <img className="suggestImg" src="/images" alt="" />
+                </div>
+                <div className="name">
+                    <span className="suggestName">Friend Name  <br /> <span className="handel">@Handel</span> </span>
+                </div>
+                <button> follow</button>
+            </div>
+            <div className="suggestFriend">
+                <div className="suggestImgContainer">
+                    <img className="suggestImg" src="/images" alt="" />
+                </div>
+                <div className="name">
+                    <span className="suggestName">Friend Name  <br /> <span className="handel">@Handel</span> </span>
+                </div>
+                <button> follow</button>
+            </div>
+            <div className="suggestFriend">
+                <div className="suggestImgContainer">
+                    <img className="suggestImg" src="/images" alt="" />
+                </div>
+                <div className="name">
+                    <span className="suggestName">Friend Name  <br /> <span className="handel">@Handel</span> </span>
+                </div>
+                <button> follow</button>
+            </div>
+            <div className="suggestFriend">
+                <div className="suggestImgContainer">
+                    <img className="suggestImg" src="/images" alt="" />
+                </div>
+                <div className="name">
+                    <span className="suggestName">Friend Name  <br /> <span className="handel">@Handel</span> </span>
+                </div>
+                <button> follow</button>
+            </div>
+            <div className="suggestFriend">
+                <div className="suggestImgContainer">
+                    <img className="suggestImg" src="/images" alt="" />
+                </div>
+                <div className="name">
+                    <span className="suggestName">Friend Name  <br /> <span className="handel">@Handel</span> </span>
+                </div>
+                <button> follow</button>
+            </div>
+            <div className="suggestFriend">
+                <div className="suggestImgContainer">
+                    <img className="suggestImg" src="/images" alt="" />
+                </div>
+                <div className="name">
+                    <span className="suggestName">Friend Name  <br /> <span className="handel">@Handel</span> </span>
+                </div>
+                <button> follow</button>
+            </div>
 
-            {users.map(user => (
-                <Link className="users__link" key={user.id} to={`/profile/${user.id}`}>
-                    <PeopleBox user={user} />
-                </Link>
-            ))}
-            <PeopleBox />
-            
         </div>
     );
 };
