+++ Client/src/Components/PersonalProfile/PersonalProfile.js
@@ -56,7 +56,7 @@ const PersonalProfile = () => {
 						<span>Add country</span>
 					</div>
                 </div>
-                <div className="PersonalProfile__country__container">        
+                <div className="PersonalProfile__country__container">
                     {countries.map(country => (
                         <PersonalProfilePost key={country._id} country={country} />
                     ))}
